# openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "The Bakers Inn API",
    "description": "REST API for The Bakers Inn bakery \u2014 public storefront reads and admin panel management.",
    "version": "1.0.0"
  },
  "paths": {
...
    "/api/v1/categories/{slug}": {
      "get": {
        "tags": [
          "categories"
        ],
        "summary": "Get one active category by slug",
        "operationId": "get_category_api_v1_categories__slug__get",
        "parameters": [
          {
            "name": "slug",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Slug"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CategoryRead"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/auth/invitations/{token}": {
      "get": {
        "tags": [
          "auth"
        ],
        "summary": "Look up a pending invitation by its token",
        "operationId": "get_invitation_api_v1_auth_invitations__token__get",
        "parameters": [
          {
            "name": "token",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Token"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/InvitationPublicInfo"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/admin/categories": {
      "post": {
        "tags": [
          "admin:categories"
        ],
        "summary": "Create a category (icon_url required)",
        "operationId": "create_category_api_v1_admin_categories_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CategoryCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CategoryRead"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/api/v1/admin/categories/reorder": {
      "patch": {
        "tags": [
          "admin:categories"
        ],
        "summary": "Bulk reorder categories",
        "operationId": "reorder_categories_api_v1_admin_categories_reorder_patch",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CategoryReorderRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/api/v1/admin/categories/{category_id}": {
      "patch": {
        "tags": [
          "admin:categories"
        ],
        "summary": "Partially update a category",
        "operationId": "update_category_api_v1_admin_categories__category_id__patch",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "category_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Category Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CategoryUpdate"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CategoryRead"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "admin:categories"
        ],
        "summary": "Delete a category (owner only)",
        "operationId": "delete_category_api_v1_admin_categories__category_id__delete",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "category_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Category Id"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/admin/invitations": {
      "get": {
        "tags": [
          "admin:invitations"
        ],
        "summary": "List invitations and join requests",
        "operationId": "list_invitations_api_v1_admin_invitations_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/InvitationRead"
                  },
                  "type": "array",
                  "title": "Response List Invitations Api V1 Admin Invitations Get"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      },
      "post": {
        "tags": [
          "admin:invitations"
        ],
        "summary": "Invite someone to the admin panel",
        "operationId": "create_invitation_api_v1_admin_invitations_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/InvitationCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/InvitationRead"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/api/v1/admin/invitations/{invitation_id}/approve": {
      "post": {
        "tags": [
          "admin:invitations"
        ],
        "summary": "Approve a self-serve join request (issues the acceptance token)",
        "operationId": "approve_invitation_api_v1_admin_invitations__invitation_id__approve_post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "invitation_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Invitation Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/InvitationRead"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/admin/invitations/{invitation_id}": {
      "delete": {
        "tags": [
          "admin:invitations"
        ],
        "summary": "Revoke a pending or requested invitation",
        "operationId": "revoke_invitation_api_v1_admin_invitations__invitation_id__delete",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "invitation_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Invitation Id"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "CategoryCreate": {
        "properties": {
          "name": {
            "type": "string",
            "maxLength": 120,
            "minLength": 1,
            "title": "Name",
            "description": "Display name of the category.",
            "examples": [
              "Cakes"
            ]
          },
          "description": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Description",
            "description": "Optional longer description shown on the storefront category page.",
            "examples": [
              "Our signature layered cakes, made fresh daily."
            ]
          },
          "display_order": {
            "type": "integer",
            "title": "Display Order",
            "description": "Sort order for the storefront category strip. Lower shows first.",
            "default": 0,
            "examples": [
              0
            ]
          },
          "icon_url": {
            "type": "string",
            "title": "Icon Url",
            "description": "URL of the category icon. Required on create \u2014 the admin workflow is icon-first, a category should never be created without one.",
            "examples": [
              "https://cdn.thebakersinn.com/icons/cakes.png"
            ]
          }
        },
        "type": "object",
        "required": [
          "name",
          "icon_url"
        ],
        "title": "CategoryCreate"
      },
      "CategoryRead": {
        "properties": {
          "name": {
            "type": "string",
            "maxLength": 120,
            "minLength": 1,
            "title": "Name",
            "description": "Display name of the category.",
            "examples": [
              "Cakes"
            ]
          },
          "description": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Description",
            "description": "Optional longer description shown on the storefront category page.",
            "examples": [
              "Our signature layered cakes, made fresh daily."
            ]
          },
          "display_order": {
            "type": "integer",
            "title": "Display Order",
            "description": "Sort order for the storefront category strip. Lower shows first.",
            "default": 0,
            "examples": [
              0
            ]
          },
          "id": {
            "type": "integer",
            "title": "Id",
            "description": "Category ID.",
            "examples": [
              1
            ]
          },
          "slug": {
            "type": "string",
            "title": "Slug",
            "description": "URL-safe slug, generated from name.",
            "examples": [
              "cakes"
            ]
          },
          "icon_url": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Icon Url",
            "description": "URL of the category icon.",
            "examples": [
              "https://cdn.thebakersinn.com/icons/cakes.png"
            ]
          },
          "is_active": {
            "type": "boolean",
            "title": "Is Active",
            "description": "Whether this category is visible on the storefront.",
            "examples": [
              true
            ]
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At",
            "description": "Creation timestamp (UTC)."
          },
          "updated_at": {
            "type": "string",
            "format": "date-time",
            "title": "Updated At",
            "description": "Last update timestamp (UTC)."
          }
        },
        "type": "object",
        "required": [
          "name",
          "id",
          "slug",
          "is_active",
          "created_at",
          "updated_at"
        ],
        "title": "CategoryRead"
      },
      "CategoryReorderItem": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id",
            "description": "Category ID to reposition.",
            "examples": [
              3
            ]
          },
          "display_order": {
            "type": "integer",
            "title": "Display Order",
            "description": "New sort position.",
            "examples": [
              2
            ]
          }
        },
        "type": "object",
        "required": [
          "id",
          "display_order"
        ],
        "title": "CategoryReorderItem"
      },
      "CategoryReorderRequest": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/CategoryReorderItem"
            },
            "type": "array",
            "title": "Items",
            "description": "Bulk list of category IDs with their new display_order.",
            "examples": [
              [
                {
                  "display_order": 0,
                  "id": 3
                },
                {
                  "display_order": 1,
                  "id": 1
                }
              ]
            ]
          }
        },
        "type": "object",
        "required": [
          "items"
        ],
        "title": "CategoryReorderRequest"
      },
      "CategoryUpdate": {
        "properties": {
          "name": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 120,
                "minLength": 1
              },
              {
                "type": "null"
              }
            ],
            "title": "Name",
            "description": "New display name.",
            "examples": [
              "Cakes"
            ]
          },
          "description": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Description",
            "description": "New description.",
            "examples": [
              "Our signature layered cakes."
            ]
          },
          "icon_url": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Icon Url",
            "description": "New icon URL.",
            "examples": [
              "https://cdn.thebakersinn.com/icons/cakes.png"
            ]
          },
          "display_order": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Display Order",
            "description": "New sort position.",
            "examples": [
              1
            ]
          },
          "is_active": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Is Active",
            "description": "Set false to hide from the public storefront without deleting.",
            "examples": [
              true
            ]
          }
        },
        "type": "object",
        "title": "CategoryUpdate",
        "description": "All fields optional for PATCH."
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "InvitationCreate": {
        "properties": {
          "email": {
            "type": "string",
            "format": "email",
            "title": "Email",
            "description": "Who to invite.",
            "examples": [
              "newbaker@thebakersinn.com"
            ]
          },
          "role": {
            "$ref": "#/components/schemas/Role",
            "description": "Role they'll get on acceptance.",
            "default": "employee",
            "examples": [
              "employee"
            ]
          }
        },
        "type": "object",
        "required": [
          "email"
        ],
        "title": "InvitationCreate",
        "description": "Owner invites a specific person."
      },
      "InvitationPublicInfo": {
        "properties": {
          "email": {
            "type": "string",
            "format": "email",
            "title": "Email",
            "description": "Email the invitation is for.",
            "examples": [
              "x@example.com"
            ]
          },
          "role": {
            "$ref": "#/components/schemas/Role",
            "description": "Role that will be granted.",
            "examples": [
              "employee"
            ]
          },
          "expires_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Expires At",
            "description": "Token expiry (UTC)."
          }
        },
        "type": "object",
        "required": [
          "email",
          "role"
        ],
        "title": "InvitationPublicInfo",
        "description": "What GET /auth/invitations/{token} returns \u2014 enough to prefill the\naccept form, nothing sensitive."
      },
      "InvitationRead": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id",
            "description": "Invitation ID.",
            "examples": [
              5
            ]
          },
          "email": {
            "type": "string",
            "format": "email",
            "title": "Email",
            "description": "Invitee email.",
            "examples": [
              "newbaker@thebakersinn.com"
            ]
          },
          "role": {
            "$ref": "#/components/schemas/Role",
            "description": "Role granted on acceptance.",
            "examples": [
              "employee"
            ]
          },
          "status": {
            "$ref": "#/components/schemas/InvitationStatus",
            "description": "requested | pending | accepted | revoked.",
            "examples": [
              "pending"
            ]
          },
          "token": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Token",
            "description": "Acceptance token. Null while status is 'requested'.",
            "examples": [
              "k3y-Vd9..."
            ]
          },
          "invited_by_id": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Invited By Id",
            "description": "ID of the owner who sent it; null for self-requests.",
            "examples": [
              1
            ]
          },
          "expires_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Expires At",
            "description": "When the token stops working (UTC)."
          },
          "accepted_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Accepted At",
            "description": "When the account was created (UTC)."
          },
          "is_expired": {
            "type": "boolean",
            "title": "Is Expired",
            "description": "True if pending but past expires_at.",
            "examples": [
              false
            ]
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At",
            "description": "Creation timestamp (UTC)."
          }
        },
        "type": "object",
        "required": [
          "id",
          "email",
          "role",
          "status",
          "is_expired",
          "created_at"
        ],
        "title": "InvitationRead",
        "description": "Admin-facing view. Includes the raw token so an owner can share the link\n(there's no email delivery configured \u2014 see CLAUDE.md \u00a76)."
      },
      "InvitationStatus": {
        "type": "string",
        "enum": [
          "requested",
          "pending",
          "accepted",
          "revoked"
        ],
        "title": "InvitationStatus"
      },
...
      "Role": {
        "type": "string",
        "enum": [
          "owner",
          "employee"
        ],
        "title": "Role"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    },
    "securitySchemes": {
      "HTTPBearer": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  }
}